outsource convert_to_str_paths
